Added pruning function
